+++ Neuro/ActivationFunctions/BipolarSigmoid.cs
@@ -4,16 +4,16 @@ namespace Neuro.ActivationFunctions
 {
     public class BipolarSigmoid : IActivationFunction
     {
-        public double Alpha { get; set; } = 2;
-        public double MinRange { get; set; } = -1;
-        public double MaxRange { get; set; } = 1;
+        public float Alpha { get; set; } = 2;
+        public float MinRange { get; set; } = -1;
+        public float MaxRange { get; set; } = 1;
         
-        public double Activation(double x)
+        public float Activation(float x)
         {
-            return (2 / (1 + Math.Exp(-Alpha * x))) - 1;
+            return (2 / (1 + (float)Math.Exp(-Alpha * x))) - 1;
         }
 
-        public double Derivative(double x)
+        public float Derivative(float x)
         {
             return Alpha * (1 - x * x) / 2;
         }
