+++ Neuro/ActivationFunctions/Sigmoid.cs
@@ -4,16 +4,16 @@ namespace Neuro.ActivationFunctions
 {
     public class Sigmoid : IActivationFunction
     {
-        public double Alpha { get; set; } = 2;
-        public double MinRange { get; set; } = 0;
-        public double MaxRange { get; set; } = 1;
+        public float Alpha { get; set; } = 2;
+        public float MinRange { get; set; } = 0;
+        public float MaxRange { get; set; } = 1;
         
-        public double Activation(double x)
+        public float Activation(float x)
         {
-            return 1 / (1 + Math.Exp(-Alpha * x));
+            return 1 / (1 + (float)Math.Exp(-Alpha * x));
         }
 
-        public double Derivative(double x)
+        public float Derivative(float x)
         {
             return  Alpha * x * (1 - x);
         }
