+++ Neuro/Domain/Layers/IConvolutionalLayer.cs
@@ -12,6 +12,6 @@ namespace Neuro.Domain.Layers
         ConvolutionalNeuron this[int index] { get; }
 
         void Init(int inputWidth, int inputHeitght);
-        double[] GetLinereOutput();
+        float[] GetLinereOutput();
     }
 }
