+++ Neuro/Layers/ConvolutionalLayer.cs
@@ -1,5 +1,4 @@
 ﻿using System.Linq;
-using System.Threading.Tasks;
 using Neuro.ActivationFunctions;
 using Neuro.Domain.Layers;
 using Neuro.Extensions;
@@ -13,7 +12,7 @@ namespace Neuro.Layers
         public LayerType Type { get; set; } = LayerType.Convolution;
         public ActivationType ActivationFunctionType { get; }
         public ConvolutionalNeuron[] Neurons { get; set; }
-        public double[][,] Outputs { get; private set; }
+        public float[][,] Outputs { get; private set; }
         public int OutputWidht { get; private set; }
         public int OutputHeight { get; private set; }
         public int KernelSize { get; private set; }
@@ -28,7 +27,7 @@ namespace Neuro.Layers
             _function = activationType.Get();
             KernelSize = kernelSize;
             Neurons = new ConvolutionalNeuron[neuronsCount];
-            Outputs = new double[neuronsCount][,];
+            Outputs = new float[neuronsCount][,];
         }
 
         public void Init(int inputWidth, int inputHeitght)
@@ -50,7 +49,7 @@ namespace Neuro.Layers
             }
         }
 
-        public double[][,] Compute(double[][,] input)
+        public float[][,] Compute(float[][,] input)
         {
             var outputs = Neurons.AsParallel().Select(n => n.Compute(input)).ToArray();
 
@@ -59,7 +58,7 @@ namespace Neuro.Layers
             return Outputs;
         }
         
-        public double[] GetLinereOutput()
+        public float[] GetLinereOutput()
         {
             return Outputs.ToLinearArray();
         }
